Add light theme styles. Clean up code.
