+++ apps/base-docs/src/components/DocChat/ChatModal.tsx
@@ -1,180 +1,17 @@
 import React, { useState, useCallback, useEffect, useRef } from 'react';
-import { fetchEventSource } from '@microsoft/fetch-event-source';
 import Modal from '../Modal';
 import ChatMessage from './ChatMessage';
 import Icon from '../Icon';
 
 import styles from './styles.module.css';
 
-// --- Get Conversation ID --- //
-async function getConversationId() {
-  let id: string | null = sessionStorage.getItem('BASE_AI_CONVERSATION_ID');
-
-  if (!id) {
-    const response = await fetch('https://api.mendable.ai/v0/newConversation', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        api_key: '0ab8984e-327c-4a8b-bea3-769ca01fac35',
-      }),
-    });
-
-    const data: { conversation_id: string } = (await response.json()) as {
-      conversation_id: string;
-    };
-
-    sessionStorage.setItem('BASE_AI_CONVERSATION_ID', data.conversation_id);
-    id = data.conversation_id;
-  }
-
-  return id;
-}
-
-// --- Submit Prompt and Stream Response --- //
-type ResponseSource = {
-  content: string;
-  data_id: string;
-  date_added: string;
-  id: number;
-  link: string;
-  manual_add: boolean;
-  relevance_score: number;
-  text: number;
-};
-
-let controller: AbortController;
-
-async function streamPromptResponse(
-  conversationId: string,
-  prompt: string,
-  setIsLoading: (isLoading: boolean) => void,
-  isGenerating: boolean,
-  setIsGenerating: (isGenerating: boolean) => void,
-  chatHistory: ChatHistoryMessage[],
-  setChatHistory: (chatHistory: ChatHistoryMessage[]) => void,
-  setConversation: (conversation: ConversationMessage[]) => void,
-  conversationContainerRef: React.RefObject<HTMLDivElement>,
-) {
-  try {
-    controller = new AbortController();
-
-    const url = 'https://api.mendable.ai/v0/mendableChat';
-
-    const data = {
-      api_key: '0ab8984e-327c-4a8b-bea3-769ca01fac35',
-      question: prompt,
-      history: chatHistory,
-      conversation_id: conversationId,
-      retriever_options: {
-        num_chunks: 4,
-      },
-    };
-
-    let responseSources: ResponseSource[];
-    let responseMessageId: unknown;
-    let fullResponse = '';
-
-    await fetchEventSource(url, {
-      method: 'POST',
-      headers: {
-        Accept: 'text/event-stream',
-        'Content-Type': 'application/json',
-      },
-      openWhenHidden: true,
-      body: JSON.stringify(data),
-      signal: controller.signal,
-      onmessage(event: unknown) {
-        const parsedData = JSON.parse(event.data);
-        const chunk: string = parsedData.chunk;
-        if (chunk === '<|source|>') {
-          responseSources = parsedData.metadata as ResponseSource[];
-          return;
-        } else if (chunk === '<|message_id|>') {
-          responseMessageId = parsedData.metadata;
-          return;
-        }
-
-        // End loading spinner and show Stop Generating button
-        if (!isGenerating) {
-          setIsLoading(false);
-          setIsGenerating(true);
-        }
-
-        // Update full response string
-        fullResponse = fullResponse.concat(chunk);
-
-        // Update rendered conversation data for current response
-        setConversation((prevState: ConversationMessage[]) => {
-          const currentResponse = prevState.slice(-1)[0];
-          const updatedResponse = {
-            type: 'response',
-            content: currentResponse.content.concat(chunk),
-          };
-
-          const newState = [...prevState.slice(0, -1), updatedResponse];
-          return newState;
-        });
-
-        // Scroll to bottom of conversation container while generating
-        conversationContainerRef.current?.scrollBy(
-          0,
-          conversationContainerRef.current.scrollHeight,
-        );
-
-        return;
-      },
-      onclose() {
-        // Add sources to current response object
-        const sourceURLs: string[] = responseSources.map((source) => source.link);
-
-        setConversation((prevState: ConversationMessage[]) => {
-          const currentResponse = prevState.slice(-1)[0];
-          const updatedResponse = {
-            ...currentResponse,
-            sources: sourceURLs,
-          };
-
-          const newState = [...prevState.slice(0, -1), updatedResponse];
-          return newState;
-        });
-
-        // Update chat history for Mendable API
-        setChatHistory((prevState: ChatHistoryMessage[]) => {
-          const currentResponse = prevState.slice(-1)[0];
-          const updatedResponse = {
-            ...currentResponse,
-            response: fullResponse,
-          };
-
-          const newState = [...prevState.slice(0, -1), updatedResponse];
-          return newState;
-        });
-
-        setIsGenerating(false);
-        return;
-      },
-      onerror(err: unknown) {
-        console.error(err);
-        return;
-      },
-    });
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-export type ConversationMessage = {
-  type: 'prompt' | 'response';
-  content: string;
-  sources?: string[];
-};
-
-type ChatHistoryMessage = {
-  prompt: string;
-  response: string;
-};
+import {
+  ConversationMessage,
+  ChatHistoryMessage,
+  getConversationId,
+  streamPromptResponse,
+  controller,
+} from './docChat';
 
 type ChatModalProps = {
   visible: boolean;
@@ -200,7 +37,9 @@ export default function ChatModal({ visible, onRequestClose }: ChatModalProps) {
     (e: React.SyntheticEvent) => {
       e.preventDefault();
 
-      if (!prompt) return;
+      if (!prompt || isLoading || isGenerating) return;
+
+      setIsLoading(true);
 
       setChatHistory((prevState: ChatHistoryMessage[]) => {
         const newState = [...prevState];
@@ -215,8 +54,6 @@ export default function ChatModal({ visible, onRequestClose }: ChatModalProps) {
         return newState;
       });
 
-      setIsLoading(true);
-
       streamPromptResponse(
         conversationId,
         prompt,
@@ -226,7 +63,6 @@ export default function ChatModal({ visible, onRequestClose }: ChatModalProps) {
         chatHistory,
         setChatHistory,
         setConversation,
-        conversationContainerRef,
       ).catch((err) => console.error(err));
 
       setPrompt('');
@@ -255,12 +91,20 @@ export default function ChatModal({ visible, onRequestClose }: ChatModalProps) {
   }, [controller]);
 
   useEffect(() => {
-    getConversationId()
-      .then((id) => {
-        setConversationId(id);
-      })
-      .catch((err) => console.error(err));
-  }, []);
+    // Only get conversation ID if modal is opened
+    if (visible) {
+      getConversationId()
+        .then((id) => {
+          setConversationId(id);
+        })
+        .catch((err) => console.error(err));
+    }
+  }, [visible]);
+
+  useEffect(() => {
+    // Scroll to bottom of conversation container
+    conversationContainerRef.current?.scrollBy(0, conversationContainerRef.current.scrollHeight);
+  }, [conversation]);
 
   return (
     <Modal visible={visible} onRequestClose={onRequestClose}>
@@ -335,17 +179,6 @@ export default function ChatModal({ visible, onRequestClose }: ChatModalProps) {
           Do not enter any sensitive information.
         </div>
       </div>
-
-      {/* <button
-        onClick={() => {
-          console.log(prompt);
-          console.log(chatHistory);
-          console.log(conversation);
-          console.log(isGenerating);
-        }}
-      >
-        Print states
-      </button> */}
     </Modal>
   );
 }
