+++ apps/base-docs/src/components/DocChat/styles.module.css
@@ -3,19 +3,38 @@
   position: fixed;
   bottom: 2rem;
   right: 2rem;
+  z-index: 100;
+}
+
+.floatingChatButton {
+  background-color: transparent;
+  border-radius: 50%;
+  cursor: pointer;
+}
+[data-theme='light'] .floatingChatButton {
+  fill: black;
+}
+[data-theme='dark'] .floatingChatButton {
+  fill: white;
 }
 
 .floatingChatButtonTooltip {
   position: absolute;
   top: -2.75rem;
   right: 0;
-  background: white;
   border-radius: 0.25rem;
   font-size: 0.75rem;
-  color: black;
   padding: 3px 10px;
   white-space: nowrap;
   animation: tooltip-fade-in 0.2s ease-in-out;
+}
+[data-theme='light'] .floatingChatButtonTooltip {
+  color: white;
+  background: black;
+}
+[data-theme='dark'] .floatingChatButtonTooltip {
+  color: black;
+  background: white;
 }
 
 .tooltipPoint {
@@ -25,15 +44,13 @@
   position: absolute;
   bottom: -0.25rem;
   right: 2rem;
-  background-color: white;
   border-radius: 1px;
 }
-
-.floatingChatButton {
-  background-color: transparent;
-  border-radius: 50%;
-  cursor: pointer;
-  fill: white;
+[data-theme='light'] .tooltipPoint {
+  background-color: black;
+}
+[data-theme='dark'] .tooltipPoint {
+  background-color: white;
 }
 
 @keyframes tooltip-fade-in {
@@ -55,7 +72,6 @@
 }
 
 .chatModalFooter {
-  height: 9.5rem;
   padding: 1.5rem;
   flex-grow: 0;
 }
@@ -77,10 +93,15 @@
   position: relative;
   overflow-y: scroll;
 }
+[data-theme='light'] .conversationContainer * {
+  fill: black;
+}
+[data-theme='dark'] .conversationContainer * {
+  fill: white;
+}
 
 .chatMessageContainer {
   width: 100%;
-  fill: white;
   white-space: pre-line;
   display: flex;
   gap: 0.75rem;
@@ -99,19 +120,30 @@
 }
 
 .chatMessageSource {
-  color: white;
   max-width: 225px;
   padding: 0.25rem 0.5rem;
   margin: 0.5rem 0.5rem 0 0;
   font-size: 0.75rem;
-  border: 1px solid rgba(138, 145, 158, 0.67);
   border-radius: 0.33rem;
   overflow-x: hidden;
   white-space: nowrap;
 }
+[data-theme='light'] .chatMessageSource {
+  color: black;
+  border: 1px solid rgba(91, 97, 110, 0.5);
+}
+[data-theme='dark'] .chatMessageSource {
+  color: white;
+  border: 1px solid rgba(138, 145, 158, 0.67);
+}
 
-[data-theme='dark'] .chatMessageDivider {
+.chatMessageDivider {
   width: 100%;
+}
+[data-theme='light'] .chatMessageDivider {
+  border-top: 1px solid rgba(91, 97, 110, 0.2);
+}
+[data-theme='dark'] .chatMessageDivider {
   border-top: 1px solid rgba(138, 145, 158, 0.2);
 }
 
@@ -120,8 +152,6 @@
   font-family: CoinbaseSans;
   padding: 0.25rem 0.5rem;
   border-radius: 0.33rem;
-  background: rgba(191, 196, 202, 0.2);
-  fill: white;
   display: inline-flex;
   gap: 0.5rem;
   align-items: center;
@@ -129,13 +159,18 @@
   transition: all 0.1s ease-in-out;
   cursor: pointer;
 }
-.resetButton:hover,
-.stopGeneratingButton:hover {
-  background: rgba(191, 196, 202, 0.25);
+[data-theme='dark'] .resetButton,
+[data-theme='dark'] .stopGeneratingButton {
+  fill: white;
+  background: rgba(58, 61, 69, 0.9);
 }
-.resetButton:active,
-.stopGeneratingButton:active {
-  background: rgba(126, 128, 130, 0.25);
+[data-theme='dark'] .resetButton:hover,
+[data-theme='dark'] .stopGeneratingButton:hover {
+  background: rgb(58, 61, 69);
+}
+[data-theme='dark'] .resetButton:active,
+[data-theme='dark'] .stopGeneratingButton:active {
+  background: rgb(50, 52, 59);
 }
 
 .resetButton {
@@ -143,6 +178,10 @@
   top: 1.4rem;
   right: 1.5rem;
   z-index: 250;
+}
+
+.stopGeneratingButton {
+  align-self: center;
 }
 
 .searchingDocsMessage {
@@ -153,20 +192,20 @@
   flex-direction: column;
   align-items: center;
   gap: 0.25rem;
-  fill: white;
-}
-
-.stopGeneratingButton {
-  align-self: center;
 }
 
 .promptForm {
-  border: 1px solid rgba(138, 145, 158, 0.67);
   padding: 1rem;
   width: 100%;
   border-radius: 0.5rem;
   display: flex;
   align-items: center;
+}
+[data-theme='light'] .promptForm {
+  border: 1px solid rgba(91, 97, 110, 0.5);
+}
+[data-theme='dark'] .promptForm {
+  border: 1px solid rgba(138, 145, 158, 0.67);
 }
 
 .promptInput {
@@ -177,12 +216,20 @@
   font-size: 1rem;
 }
 
-.promptInput::placeholder {
+[data-theme='light'] .promptInput::placeholder {
+  color: rgba(91, 97, 110, 0.5);
+}
+[data-theme='dark'] .promptInput::placeholder {
   color: rgba(138, 145, 158, 0.67);
 }
 
 .promptInputIcon {
   display: flex;
+}
+[data-theme='light'] .promptInputIcon {
+  fill: rgba(91, 97, 110, 0.5);
+}
+[data-theme='dark'] .promptInputIcon {
   fill: rgba(191, 196, 202, 0.67);
 }
 
@@ -192,10 +239,16 @@
   appearance: none;
   background: transparent;
 }
-.submitPromptButton:hover {
+[data-theme='light'] .submitPromptButton:hover {
+  fill: rgba(91, 97, 110, 0.67);
+}
+[data-theme='light'] .submitPromptButton:active {
+  fill: rgba(75, 79, 90, 0.67);
+}
+[data-theme='dark'] .submitPromptButton:hover {
   fill: rgba(191, 196, 202, 0.8);
 }
-.submitPromptButton:active {
+[data-theme='dark'] .submitPromptButton:active {
   fill: rgba(164, 168, 172, 0.8);
 }
 
@@ -222,10 +275,14 @@
   }
 }
 
-@media (max-width: 997px) {
+@media (max-width: 832px) {
   .floatingChatButtonContainer {
     bottom: 1rem;
     right: 1rem;
+  }
+
+  .conversationContainer {
+    max-width: 100%;
   }
 }
